requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-dotenv==1.1.1
PyMySQL==1.1.1
oracledb==3.2.0
PyJWT==2.10.1
argon2-cffi==25.1.0
httpx==0.28.1
pytest==8.4.1
